splitted components folder into one for lists and displaying the tasks and one for the editing components
