+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import "./App.css";
 
-import { DisplayTask } from "./components/DisplayTask";
+import { DisplayTask } from "./list-components/DisplayTask";
 import { Task } from "./interfaces/task";
-import { ChangeRole } from "./components/selectuser";
+import { ChangeRole } from "./list-components/selectuser";
 import { User } from "./interfaces/user";
 
 const TASK: Task[] =
